+++ hooks/pre_commit/default_branch.rb
@@ -4,6 +4,8 @@
 
 branch = `git rev-parse --abbrev-ref HEAD`.strip
 if %w[master main].include?(branch)
+  warn ''
   warn "Commits on '#{branch}' are not allowed. Create a feature branch."
+  warn ''
   exit 1
 end